split up flask files
