+++ flask_app/static/images/logo_small.png
