+++ flask_app/templates/sensor.html
